+++ src/main/java/com/fourirbnb/common/security/Role.java
@@ -1,0 +1,5 @@
+package com.fourirbnb.common.security;
+
+public enum Role {
+  MASTER, MANAGER, HOST, CUSTOMER
+}